+++ frontend/src/app/notifications/configure/checkbox.tsx
@@ -17,7 +17,7 @@ const Checkbox = ({ checked, onChange }: CheckboxProps) => {
       <span
         className={styles.checked}
         style={{
-          backgroundColor: checked ? "#1ed760" : "transparent",
+          backgroundColor: checked ? "#ae96fd" : "transparent",
           border: checked ? "none" : "1px solid #727272",
         }}
       ></span>
